Handle a "Not charging" state.
In two-battery laptops, sometimes one of the two batteries may now
apparently show up as "Not charging". This can happen both in the case
of the other battery providing power (i.e., showing as "Discharging",
as well as the case of the other battery being charged (i.e., there not
being enough power, or the battery being close enough to fully loaded,
so that the one shows as "Charging" and the other as "Discharging".

We however still need to consider a case of "Not charging" for a
single-battery system (which likely means the system is plugged in and
not fully loaded, but no power being routed to the battery by the
firmware to prevent damage to the battery).

As such, always assign any value of "Charging" or "Discharging" to the
$state variable, but assign any other values only if the state variable
is othewise undefined.

Closes: gh-3